Bump Go to 1.23 (#487)
Go 1.21 is no longer maintained.

+++ cmd/server/main.go
@@ -238,7 +238,8 @@ func main() {
 	}
 
 	zc := zap.NewProductionConfig()
-	zc.Level = zap.NewAtomicLevelAt(zapcore.Level(-1 * config.LogVerbosity))
+	// #nosec G115: when parsing config, we ensure that LogVerbosity is >= 0 and < 128.
+	zc.Level = zap.NewAtomicLevelAt(zapcore.Level(-1 * int8(config.LogVerbosity)))
 	zc.DisableStacktrace = true
 	zapLog, err := zc.Build()
 	if err != nil {
